moved limit as table function which allows flexibility and cleaner DSL

+++ src/main/kotlin/com/github/dataanon/utils/ProgressBarGenerator.kt
@@ -3,13 +3,13 @@ package com.github.dataanon.utils
 import me.tongfei.progressbar.ProgressBar
 import me.tongfei.progressbar.ProgressBarStyle
 
-class ProgressBarGenerator(private val progressBarEnabled: Boolean = true, taskName: String, initialMaxFn: () -> Long) {
+class ProgressBarGenerator(private val progressBarEnabled: Boolean = true, taskName: String, initialMaxFn: () -> Int) {
 
     private lateinit var pb: ProgressBar
 
     init {
         if (progressBarEnabled)
-            pb = ProgressBar(taskName, initialMaxFn(), ProgressBarStyle.ASCII)
+            pb = ProgressBar(taskName, initialMaxFn().toLong(), ProgressBarStyle.ASCII)
     }
 
     fun start() {
